requirements.txt
requests
pandas
SQLAlchemy
python-dotenv
pytest
logging